Put ViewPagerAdapter into separate file
